Split component to separate file.
